requirements.txt
PyPDF2
tqdm
pdfkit
pymupdf
google-generativeai
requests
beautifulsoup4
python-dotenv 